+++ packages/css-should/src/helpers/pluck.js
@@ -1,0 +1,5 @@
+export default function (property, arr) {
+  return arr
+    .map((obj) => obj[property])
+    .filter((x) => !!x)
+}