+++ packages/css-should/src/middleware/not-handled.js
@@ -1,0 +1,2 @@
+export const test = ({ test, fail }) =>
+  fail(test, new Error('Unhandled test with "' + test.axis + '" axis.'))